requirements.txt
Flask
pytesseract
Pillow
joblib
numpy
scikit-learn
gunicorn